misc/HtmlTestSuite/results/test-0399.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml">
<head>
<meta charset="utf-8"/>
<title>CSS Backgrounds and Borders Module Level 3: parsing border with valid values</title>
<link href="https://drafts.csswg.org/css-backgrounds/#border-shorthands" rel="help"/>
<meta content="border supports the full grammar '&lt;line-width&gt; || &lt;line-style&gt; || &lt;color&gt;'." name="assert"/>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<script src="/css/support/parsing-testcommon.js"/>
</head>
<body>
<script>
test_valid_value("border", "1px dotted red");
test_valid_value("border", "green double thin", "thin double green");

test_valid_value("border-top", "thin", ["thin", "thin none"]);
test_valid_value("border-right", "double", ["double", "medium double"]);
test_valid_value("border-bottom", "green", ["green", "medium none green"]);
test_valid_value("border-left", "1px dotted red");
</script>
</body>
</html>
